+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Added basic naming support for M3 and M3-referenced files.
 - Added naming of icon names from the web API.
 - Added support for new M3 instance chunk version (fixes M3 parsing).
-- Added setting for turning off BLP transparency in previews.
+- Added button/setting for turning off BLP transparency in previews.
 - Updated M3 modelviewer with basic texture loading.
 - Updated pagination on builds page to use input pagination.
 - Updated TACT key metadata.